+++ buzz/__init__.py
@@ -4,6 +4,7 @@ from buzz.tools import (
     check_expressions,
     get_traceback,
     handle_errors,
+    enforce_defined,
     reformat_exception,
     require_condition,
 )
@@ -14,6 +15,7 @@ __all__ = [
     "check_expressions",
     "handle_errors",
     "require_condition",
+    "enforce_defined",
     "reformat_exception",
     "get_traceback",
     "DoExceptParams",