+++ CHANGELOG.rst
@@ -7,9 +7,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+v3.2.0 - 2022-06-20
+-------------------
+* Added ``enforce_defined()`` function with tests and docs.
+
 v3.1.1 - 2022-04-05
 -------------------
-* Added `raise from` in handle_errors (how did I miss this for so long?)
+* Added ``raise from`` in handle_errors (how did I miss this for so long?)
 
 v3.1.0 - 2022-03-23
 -------------------